fix(tests): tiny_db fixture 提升到 conftest — main_routes 同摆脱 data/ 依赖
- conftest.tiny_db: tmp 最小 ipinfo_lite 库 + env 继承给 spawn 子进程 +
  load_db 钉死 no-op (测试内双载不再双开 LMDB)
- test_main_routes: autouse 挂 tiny_db, 否则 CI 冷启动查询门 503
- test_batch_pool: 本地 fixture 改用共享版
本地空 DATA_DIR 模拟: core 47 文件仅剩 api_tasks(需真网络, CI 已验过)
