+++ backend/tests/core/test_batch_pool.py
@@ -1,27 +1,10 @@
 """S1 batch process-pool: layout sizing + fan-out."""
 import pytest
 
-# Module under test (created in this task)
+# Module under test (created in this task); tiny_db fixture 来自 tests/conftest.py
 from ipdb import _batch_pool
 from ipdb import _registry
 from concurrent.futures.process import BrokenProcessPool
-
-
-@pytest.fixture
-def tiny_db(tmp_path, monkeypatch):
-    """最小库: tmp 下建一个 ipinfo_lite LMDB。
-    进程内 monkeypatch 换 _sources; spawn 子进程靠 IP_RADAR_DATA_DIR 环境继承
-    (monkeypatch 过不了进程边界)。CI 干净检出没有 data/, 不建库则 Database not loaded。"""
-    from ipdb._sources._lmdb import rebuild_lmdb
-    envs = []
-    rebuild_lmdb([
-        ("8.8.8.0/24", {"country_code": "US", "_net": "8.8.8.0/24", "has_asn": False}),
-        ("1.1.1.0/24", {"country_code": "AU", "_net": "1.1.1.0/24", "has_asn": False}),
-    ], tmp_path / "ipinfo_lite.csv.lmdb", envs.append)
-    envs[0].close()  # py-lmdb 同路径双开禁止, rebuild 句柄先关
-    monkeypatch.setenv("IP_RADAR_DATA_DIR", str(tmp_path))
-    monkeypatch.setattr(_registry, "_sources", _registry._discover_sources(tmp_path))
-    _registry.load_db()  # fixture 统一 preload — load() 不幂等, 测试里不再重复调
 
 
 @pytest.mark.parametrize("cpu, ram_mb, expected", [
